removed WParsedInt "morethanfour", "plusone","plustwo","plusthree" added 2 things first plus/plusend, minus/minusend written many ways damage:type:*:myhandminus2minusend would deal damage equal to the cards in my hand -2 damage:type:creature:mybattlefieldplustype:creature:opponentbattlefieldplusend damage equal to the number of creatures in my battlefield plus the number of creatures in opponents battlefield.
now there are special cases where we need to know the value BEFORE applying multipliers. life: is one of them. I added "math/mathend"

auto=@each opponent upkeep:target(opponent) life:-mathtype:*:myhandminustype:*:opponenthandminusendmathend opponent
in the example above, without "math" the "-" would be used for the first section of the minus parsed int, instead of being applied as a multiplier for the total of the number.

this is extremely flexible, and can contain any Wparsed int in both parts, you can add or subtract and use the differences between to values as the value of the parsedInt.