Replace 'AE' in card names with 'Ae'
This is caused by a retroactive change described at https://markrosewater.tumblr.com/post/144471532728/.
In case link stops working, here is its content quoted:

"knight-otu asked: This is a bit weird, but I notice it's the Aether Revolt, rather than the Æther Revolt.
Is that just for the ease of typing the set name?

The ligature is dead! (And I’m proud to say, I had a huge hand in killing it.)
The change will be retroactive. Cards with the Ae ligature will now just be an “A” and an “e”.
Hallelujah!"